+++ packages/api/src/Domain/Response/SharedVault/GetSharedVaultsResponse.ts
@@ -2,4 +2,8 @@ import { SharedVaultServerHash } from '@standardnotes/responses'
 
 export type GetSharedVaultsResponse = {
   sharedVaults: SharedVaultServerHash[]
+  designatedSurvivors: Array<{
+    userUuid: string
+    sharedVaultUuid: string
+  }>
 }