+++ packages/models/src/Domain/Syncable/VaultListing/VaultListingSharingInfo.ts
@@ -2,4 +2,5 @@ export type VaultListingSharingInfo = {
   sharedVaultUuid: string
   ownerUserUuid: string
   fileBytesUsed: number
+  designatedSurvivor: string | null
 }