+++ packages/web/src/javascripts/Controllers/LinkingController.tsx
@@ -213,7 +213,10 @@ export class LinkingController extends AbstractViewController implements Interna
           const noteVault = this.vaults.getItemVault(note)
           const fileVault = this.vaults.getItemVault(updatedFile)
           if (noteVault && !fileVault) {
-            await this.vaults.moveItemToVault(noteVault, file)
+            const result = await this.vaults.moveItemToVault(noteVault, file)
+            if (result.isFailed()) {
+              console.error(result.getError())
+            }
           }
         }
       }
@@ -305,10 +308,12 @@ export class LinkingController extends AbstractViewController implements Interna
       const itemVault = this.vaults.getItemVault(activeItem)
 
       if (itemVault) {
-        const movedTag = await this.vaults.moveItemToVault(itemVault, newTag)
-        if (!movedTag) {
-          throw new Error('Failed to move tag to item vault')
+        const movedTagOrError = await this.vaults.moveItemToVault(itemVault, newTag)
+        if (movedTagOrError.isFailed()) {
+          throw new Error(`Failed to move tag to item vault: ${movedTagOrError.getError()}`)
         }
+        const movedTag = movedTagOrError.getValue()
+
         await this.addTagToItem(movedTag as SNTag, activeItem)
       } else {
         await this.addTagToItem(newTag, activeItem)