+++ packages/web/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.173.15](https://github.com/standardnotes/app/compare/@standardnotes/web@3.173.14...@standardnotes/web@3.173.15) (2023-09-22)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.173.14](https://github.com/standardnotes/app/compare/@standardnotes/web@3.173.13...@standardnotes/web@3.173.14) (2023-09-22)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.173.13](https://github.com/standardnotes/app/compare/@standardnotes/web@3.173.12...@standardnotes/web@3.173.13) (2023-09-22)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.173.12](https://github.com/standardnotes/app/compare/@standardnotes/web@3.173.11...@standardnotes/web@3.173.12) (2023-09-21)
 
 **Note:** Version bump only for package @standardnotes/web